project updated to Visual Studio 2013;
rebuild udis; updated project options /W4 /WX; readme updated; build
script updated; code style fixes; strings routines not inlined

+++ src/pnptable.cpp
@@ -147,7 +147,7 @@ EXT_COMMAND(wa_pnptable, "Output kernel-mode nt!PlugPlayHandlerTable", "") {
     display->PrintHeader();
 
     try {
-        for ( int i = 0; i < 0x100; i++ ) {
+        for ( unsigned long i = 0; i < 0x100; i++ ) {
             ExtRemoteData pnp_table_entry_class(offset + i * size, static_cast<ULONG>(sizeof(unsigned __int32)));
 
             if ( pnp_table_entry_class.GetUlong() != i )    // check PnpControlClass
